Add files via upload
TRLG: Fragile blind quad watermarking for image tamper detection and recovery by providing compact digests with optimized quality using LWT and GA

In this paper, an efficient fragile blind quad watermarking scheme, named TRLG, is proposed for image tamper detection and recovery based on lifting wavelet transform and genetic algorithm. TRLG generates four compact digests with super quality based on lifting wavelet transform and halftoning technique by distinguishing the types of image blocks. In this way, for each 2  ×  2 non-overlapping block, four chances for recovering the destroyed blocks are created. A special parameter estimation technique based on the genetic algorithm is performed to improve and optimize the quality of digests and the watermarked image. Furthermore, the Chebyshev System is used to determine the mapping block for embedding, encrypting, and shuffling the information. To improve the recovery rate, two techniques called Mirror-aside and Partner-block are proposed. Some experiments are conducted to prove the superiority of TRLG in terms of quality of the watermarked and recovered images, tamper localization, and security compared with the state-of-the-art methods. The results indicate that the average values of PSNR and SSIM of the watermarked image are about 46 dB and 1, respectively. Also, the average values of PSNR and SSIM for several recovered images that were destroyed about 90% reached to 24 dB and 0.86, respectively.

http://www.sciencedirect.com/science/article/pii/S0020025519301707